correct printing of size_t in 32 bit mode
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/mpc/trunk@497 211d60ee-9f03-0410-a15a-8952a2c7a4e4

+++ tests/tio_str.c
@@ -102,7 +102,7 @@ check_file (const char* file_name)
                    inex, expected_inex);
           if (size !=  expected_size)
             printf ("     got size: %lu\nexpected size: %lu\n     ",
-                    size, expected_size);
+                    (unsigned long int) size, (unsigned long int) expected_size);
           OUT (got);
           OUT (expected);
